added: `UnqualifiedType` and `Declaration` methods.
These methods are used to extract the `unqualified` type of the field under
analysis and the name of the field as it appears when declared, respectively.
These are useful and needed when matching for primitive types and when
printing to screen e.g. when using `-optimize`.

Note that `UnqualifiedType` actually returns the type with no qualifiers
except signedness and `long`-ness ones.