changed: refactoring `extract.go` and `stropt.go` post-parsing changes.
Moved most of the parsing code to `parse.go` as parte of the parsing rework
action. The field-kind discrimination action is now performed based on the
underlying type of the field, which has become an interface.

Data extraction is now performed on the translation unit, only when
encountering a declarator which has a struct, union or enum as the type
it is descripting.

Removing debug mode and updating the prints to use the new `Field` APIs.